security: land reviewed WillBeOut remediation stack

+++ attendees.py
@@ -1,33 +1,35 @@
 import tornado.web
-import tornado.auth
 import base
 import json
 
 
-class Attend(base.BaseHandler, tornado.auth.FacebookGraphMixin):
+class Attend(base.BaseHandler):
     @tornado.web.authenticated
-    def post(self):
+    async def post(self):
         _event_id = self.get_int_argument('event_id')
+        await self.require_event_access(_event_id)
         _user_id = self.get_current_user()['id']
         self.db.execute(
             "INSERT INTO willbeout_attendees (user_id, event_id) VALUES (%s, %s)",
             _user_id, _event_id)
         self.redirect('/event?event_id=' + str(_event_id))
 
-class AttendNo(base.BaseHandler, tornado.auth.FacebookGraphMixin):
+class AttendNo(base.BaseHandler):
     @tornado.web.authenticated
-    def post(self):
+    async def post(self):
         _event_id = self.get_int_argument('event_id')
+        await self.require_event_access(_event_id)
         _user_id = self.get_current_user()['id']
         self.db.execute(
             "DELETE FROM willbeout_attendees WHERE event_id = %s AND user_id  = %s",
             _event_id, _user_id,)
         self.redirect('/event?event_id=' + str(_event_id))
 
-class AttendData(base.BaseHandler, tornado.auth.FacebookGraphMixin):
+class AttendData(base.BaseHandler):
     @tornado.web.authenticated
-    def get(self):
+    async def get(self):
         _event_id = self.get_int_argument('event_id')
+        await self.require_event_access(_event_id)
         a = self.db.query(
             "SELECT * FROM willbeout_attendees WHERE event_id = %s GROUP BY user_id",
             _event_id)
