+++ events.py
@@ -1,56 +1,32 @@
-import tornado.auth
 import tornado.web
 import base
 import json
-import urllib
+from urllib.parse import unquote
 
 
-class EventHandler(base.BaseHandler, tornado.auth.FacebookGraphMixin):
+class EventHandler(base.BaseHandler):
     @tornado.web.authenticated
-    @tornado.web.asynchronous
-    def get(self):
-        # access trigger
-        self.access = 0
+    async def get(self):
         _eventid = self.get_int_argument('event_id')
-        _id = self.get_current_user()['id']
-        # get owner_id
-        print 'checkpoint 1'
-        self.event = self.db.get(
-            "SELECT * FROM willbeout_events WHERE id = %s", _eventid)
-        if not self.event:
-            raise tornado.web.HTTPError(404)
+        self.event = await self.require_event_access(_eventid)
+        _user_id = self.get_current_user()['id']
         self.suggest = self.db.query("""select a.id, a.event_id, a.address, a.city, a.name, a.url, a.user_id, a.user_name, count(b.suggestion_id) as friends from willbeout_suggest as a
-        LEFT JOIN willbeout_votes as b ON a.id = b.suggestion_id
+        LEFT JOIN willbeout_votes as b ON a.id = b.suggestion_id AND a.event_id = b.event_id
         WHERE a.event_id = %s
         GROUP BY a.id ORDER BY friends DESC;""", _eventid)
         self.votes = self.db.query("""select suggestion_id from willbeout_votes where event_id = %s and user_id = %s group by suggestion_id;
-        """, _eventid, int(_id))
+        """, _eventid, int(_user_id))
 
-        if str(self.event['userid']).strip('L') == _id:
-            self.access = 1
-        self.facebook_request("/me/friends/" + str(self.event['userid']), self
-            ._go, access_token=self.current_user["access_token"])
+        self.render(
+            "event.html",
+            event=self.event,
+            suggestions=self.suggest,
+            votes=self.votes,
+            x=self.xsrf_form_html(),
+        )
 
-    def _friendship_visible(self, streams):
-        if isinstance(streams, dict):
-            data = streams.get("data")
-            return isinstance(data, list) and len(data) > 0
-        return bool(streams)
-
-    def _go(self, streams):
-        if self.access != 1 and not self._friendship_visible(streams):
-            raise tornado.web.HTTPError(403)
-
-        try:
-            self.render('event.html', event=self.event, suggestions=self.
-                        suggest, votes=self.votes, x=self.xsrf_form_html())
-        except:
-            self.redirect('/problem')
-            pass
-
-class EventsHandler(base.BaseHandler, tornado.auth.FacebookGraphMixin):
+class EventsHandler(base.BaseHandler):
     @tornado.web.authenticated
-    @tornado.web.asynchronous
     def get(self):
         x = self.xsrf_form_html()
         _id = self.get_current_user()['id']
@@ -59,7 +35,6 @@ class EventsHandler(base.BaseHandler, tornado.auth.FacebookGraphMixin):
         self.render("events.html", x=x, events=events)
 
     @tornado.web.authenticated
-    @tornado.web.asynchronous
     def post(self):
         _id = self.get_current_user()['id']
         _place = self.get_argument('place')
@@ -71,27 +46,45 @@ class EventsHandler(base.BaseHandler, tornado.auth.FacebookGraphMixin):
         self.redirect('/events')
 
 
-class TimeHandler(base.BaseHandler, tornado.auth.FacebookGraphMixin):
+class TimeHandler(base.BaseHandler):
+    @staticmethod
+    def parse_available_times(value):
+        try:
+            if value == "":
+                return []
+            tokens = unquote(value).split(',')
+            if any(not token for token in tokens):
+                raise ValueError
+            times = [int(token) for token in tokens]
+            if len(times) != len(set(times)) or any(time < 9 or time > 24 for time in times):
+                raise ValueError
+            return times
+        except (TypeError, ValueError):
+            raise tornado.web.HTTPError(400)
+
     @tornado.web.authenticated
-    def post(self):
+    async def post(self):
         _user_id = self.get_current_user()['id']
         _user_name = self.get_current_user()['name']
         _event_id = self.get_int_argument('event_id')
-        _times = self.get_argument('availabletimes')
-        self.db.execute(
+        await self.require_event_access(_event_id)
+        _times = self.parse_available_times(self.get_argument('availabletimes'))
+        statements = [(
             "DELETE FROM willbeout_availability WHERE event_id = %s and user_id = %s",
-            _event_id, int(_user_id))
-
-        for i in urllib.unquote(_times).split(','):
-            self.db.execute(
-                """INSERT INTO willbeout_availability (user_id, user_name, time, event_id) VALUES (%s, %s, %s, %s)""",
-                int(_user_id), _user_name, int(i), _event_id)
+            (_event_id, int(_user_id)),
+        )]
+        statements.extend((
+            """INSERT INTO willbeout_availability (user_id, user_name, time, event_id) VALUES (%s, %s, %s, %s)""",
+            (int(_user_id), _user_name, i, _event_id),
+        ) for i in _times)
+        self.db.execute_transaction("willbeout_availability", statements)
         self.redirect('/event?event_id=' + str(_event_id))
 
     @tornado.web.authenticated
-    def get(self):
+    async def get(self):
         _json = []
         _event_id = self.get_int_argument('event_id')
+        await self.require_event_access(_event_id)
         for i in self.db.query(
             'SELECT * FROM willbeout_availability WHERE event_id = %s ORDER BY time',
                 _event_id):
